Use packages to store information about each variant.
